pyproject.toml
[tool.poetry]
name = "cicd_pipeline_practice"
version = "0.1.0"
description = "CI/CD Pipeline built on CDK Pipelines"
authors = ["Blake Enyart <blake.enyart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.87.1"
"aws-cdk.aws-codepipeline" = "^1.87.1"
"aws-cdk.aws-codepipeline-actions" = "^1.87.1"
"aws-cdk.pipelines" = "^1.87.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0","setuptools>=41.0", "setuptools-scm", "wheel"]
build-backend = "poetry.core.masonry.api"
